refactor(cabal-install): separate GenericReadyPackage from ReadyPackage
