+++ cabal-install/cabal-install.cabal
@@ -222,6 +222,7 @@ library
         Distribution.Client.Types.ConfiguredId
         Distribution.Client.Types.ConfiguredPackage
         Distribution.Client.Types.Credentials
+        Distribution.Client.Types.GenericReadyPackage
         Distribution.Client.Types.InstallMethod
         Distribution.Client.Types.OverwritePolicy
         Distribution.Client.Types.PackageLocation
